+++ hutool-core/src/main/java/cn/hutool/core/io/resource/MultiFileResource.java
@@ -3,6 +3,13 @@ package cn.hutool.core.io.resource;
 import java.io.File;
 import java.util.Collection;
 
+/**
+ * 多文件组合资源<br>
+ * 此资源为一个利用游标自循环资源，只有调用{@link #next()} 方法才会获取下一个资源，使用完毕后调用{@link #reset()}方法重置游标
+ * 
+ * @author looly
+ *
+ */
 public class MultiFileResource extends MultiResource{
 	
 	/**