+++ hutool-extra/src/main/java/cn/hutool/extra/template/Engine.java
@@ -8,13 +8,6 @@ package cn.hutool.extra.template;
 public interface Engine {
 	
 	/**
-	 * 获取模板引擎名
-	 * @return 模板引擎名
-	 * @since 4.1.3
-	 */
-	String getName();
-
-	/**
 	 * 获取模板
 	 * 
 	 * @param resource 资源，根据事先不同，此资源可以是模板本身，也可以是模板的相对路径