+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * 【core】         CharUtil增加isFileSeparator方法
 * 【core】         增加UUID类，提升Simple模式下性能
 * 【poi】           ExcelUtil增加setStyleSet方法，修改write逻辑，对于单列数据输出，而非忽略（感谢@【宁波】mojie126）
+* 【core】         新增WebAppResource类
+* 【extra】        新增Thymeleaf模板支持
 
 ### Bug修复